Add files via upload
Liver diseases are fast being recognized as public health priorities in India. The burden of liver disease in India is significant because it alone contributed to 18.3% of the two million global liver disease–related deaths in 2015. Through Machine Learning and Hypothesis Testing we tried to predict Liver Patients or not and whether or not it affected men and women equally. It would help doctors and medical practitioners or the patient themselves to determine early whether they are suffering from Liver disease.